support for buffers as front and back pdfs
If a buffer is passed in as a value for front or back, skip creating
a readable stream. When the time comes to encode the form submission,
if any buffers are found, set the filename to be a pdf so that the
content type is correctly set for the stream.